refactor(react-bloc): ♻️ rename file name from app-provider to root-provider
